feat:卡片上的评论的编辑和删除功能只有创建者可以使用

+++ imports/client/views/project/ProjectWindow/ProjectBord/projectActive.jsx
@@ -32,20 +32,24 @@ class ProjectMembers extends Component {
             func();
         }
     }
-    showCreadite = (id, item) => {
-        this.setState({
-            [`shownCreadite${id}`]: true,
-            changeMark: item,
-        });
+    showCreadite = (id, item, userId) => {
+        if (Meteor.userId() === userId) {
+            this.setState({
+                [`shownCreadite${id}`]: true,
+                changeMark: item,
+            });
+        }
     }
     handleChangeTry=(name, e) => {
         const newState = {};
         newState[name] = e.target.value;
         this.setState(newState);
     }
-    handleRemove = (id) => {
+    handleRemove = (id, userId) => {
         // this.deleteActive(id);
-        feedback.dealDelete('提示', '确定要删除该评论么?', () => this.deleteActive(id));
+        if (Meteor.userId() === userId) {
+            feedback.dealDelete('提示', '确定要删除该评论么?', () => this.deleteActive(id));
+        }
     }
     deleteActive = (id) => {
         Meteor.call(
@@ -132,9 +136,9 @@ class ProjectMembers extends Component {
                                     <p>{MarkValue.content}</p>
                                     <span>{format('yyyy-MM-dd', MarkValue.createTime)}</span>
                                     <span>--</span>
-                                    <a onClick={() => this.showCreadite(MarkValue._id, MarkValue.content)}>编辑</a>
+                                    <a onClick={() => this.showCreadite(MarkValue._id, MarkValue.content, MarkValue.userId)}>编辑</a>
                                     <span>--</span>
-                                    <span onClick={() => this.handleRemove(MarkValue._id)} >
+                                    <span onClick={() => this.handleRemove(MarkValue._id, MarkValue.userId)} >
                                     删除
                                     </span>
                                 </div>
